requirements.txt
appdirs==1.4.4
asttokens==2.4.1
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
clarabel==0.7.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cvxpy==1.4.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
ecos==2.0.13
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.50.0
frozendict==2.4.0
html5lib==1.1
idna==3.6
importlib_metadata==7.0.2
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.24.4
openpyxl==3.1.2
osqp==0.6.5
packaging==24.0
pandas==2.0.3
parso==0.8.3
peewee==3.17.1
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyaml==23.12.0
pybind11==2.11.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qdldl==0.1.7.post0
requests==2.31.0
scipy==1.11.4
scs==3.2.4.post1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.37
zipp==3.18.1
